wip: remove 202209 and some 202304 files
